backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
httpx==0.25.2
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.6
flower==2.0.1
google-generativeai==0.3.1
stripe==7.6.0
pyjwt==2.8.0
bcrypt==4.1.2
tenacity==8.2.3
email-validator==2.1.0
loguru==0.7.2
jinja2==3.1.2
faker==20.1.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
lxml==4.9.3
pandas==2.1.3
openpyxl==3.1.2
phonenumbers==8.13.26
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
python-magic==0.4.27
python-slugify==8.0.1
shortuuid==1.0.11
geo-py==1.0.0
country_converter==1.1.1
holidays==0.36
APScheduler==3.10.4
pytz==2023.3.post1
itsdangerous==2.1.2
starlette==0.32.0.post1
anyio==3.7.1
h11==0.14.0
greenlet==3.0.1
dnspython==2.4.2
urllib3==2.1.0
watchfiles==0.21.0
websockets==12.0
python-json-logger==2.0.7
sentry-sdk==1.39.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp==1.21.0
pyyaml==6.0.1
regex==2023.10.3
chardet==5.2.0
idna==3.4
certifi==2023.11.17
msgpack==1.0.7
yarl==1.9.3
frozenlist==1.4.0
multidict==6.0.4
aiosignal==1.3.1
attrs==23.1.0
wrapt==1.16.0
packaging==23.2
portalocker==2.8.2
pywin32==306; sys_platform == 'win32'
colorama==0.4.6
tqdm==4.66.1
click==8.1.7
typer==0.9.0
rich==13.7.0
markdown==3.5.1
bleach==6.1.0
nh3==0.2.14
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
dirty-json==1.0.8
inflect==7.0.0
humanize==4.9.0
arrow==1.3.0
pendulum==2.1.2
dateparser==1.2.0
python-dateutil==2.8.2
backoff==2.2.1
retrying==1.3.4
structlog==23.2.0
better-exceptions==0.3.3
devtools==0.12.2
icecream==2.1.3
stackprinter==0.2.10
objgraph==3.6.0
fire==0.5.0
sh==2.0.6
plumbum==1.8.2
delegator.py==0.1.1
invoke==2.2.0
fabric==3.2.0
paramiko==3.4.0
scp==0.14.5
sshtunnel==0.4.0
pycryptodome==3.19.0
cryptography==41.0.7
pyopenssl==23.3.0
service-identity==23.1.0
trustme==1.1.0
hyperlink==21.0.0
idna-ssl==1.1.0
w3lib==2.1.2
parsel==1.8.1
cssselect==1.2.0
jmespath==1.0.1
itemadapter==0.8.0
itemloaders==1.1.0
scrapinghub==2.4.0
scrapy==2.11.0 ; python_version >= "3.8"
scrapy-splash==0.9.0
scrapy-zzy==0.0.1
scrapy-fake-useragent==1.4.4
scrapy-selenium==0.0.7
scrapy-playwright==0.0.31
undetected-chromedriver==3.5.3
fake-useragent==1.4.0
requests-html==0.10.0
requests-cache==1.1.1
requests-toolbelt==1.0.0
urllib3-secure-extra==0.1.0
tldextract==5.1.1
publicsuffix2==2.20191221
whois==0.9.27
python-whois==0.8.0
dnspython==2.4.2
aiodns==3.1.1
pycares==4.4.0
cchardet==2.1.7
aiofiles==23.2.1
aiohttp==3.9.1
aiohttp-socks==0.8.4
aiohttp-retry==2.8.3
async-timeout==4.0.3
ujson==5.9.0
orjson==3.9.10
rapidfuzz==3.5.2
fuzzywuzzy==0.18.0
Levenshtein==0.23.0
textblob==0.17.1
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.2
scipy==1.11.4
modin[ray]==0.25.0
pyspark==3.5.0
